requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.8.1=py38h96a0964_0
aiosignal=1.2.0=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=21.4.0=pyhd8ed1ab_0
blas=1.0=mkl
bottleneck=1.3.4=py38h67323c0_0
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py38h96a0964_1003
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.3.18=hecd8cb5_0
cachetools=5.0.0=pyhd8ed1ab_0
certifi=2021.10.8=py38hecd8cb5_2
cffi=1.15.0=py38hc55c11b_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
click=7.0=py_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=36.0.2=py38h826b3c8_0
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
frozenlist=1.3.0=py38h96a0964_0
giflib=5.2.1=haf1e3a3_0
google-api-core=2.5.0=pyhd8ed1ab_0
google-api-core-grpc=2.5.0=hd8ed1ab_0
google-auth=2.6.2=pyh6c4a22f_0
google-cloud-core=2.2.2=pyh6c4a22f_0
google-cloud-firestore=2.3.1=pyhd8ed1ab_0
googleapis-common-protos=1.56.0=py38h50d1736_0
grpcio=1.42.0=py38ha29bfda_0
grpcio-status=1.41.1=pyhd8ed1ab_0
idna=3.3=pyhd8ed1ab_0
intel-openmp=2021.4.0=hecd8cb5_3538
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
kiwisolver=1.3.2=py38he9d5cce_0
lcms2=2.12=hf1fd2bf_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.15.8=hcf210ce_1
libtiff=4.2.0=h87d7836_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
llvm-openmp=12.0.0=h0dcd299_1
lz4-c=1.9.3=h23ab428_1
matplotlib=3.5.1=py38hecd8cb5_1
matplotlib-base=3.5.1=py38hfb0c5b7_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
multidict=6.0.2=py38h96a0964_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_2
numexpr=2.8.1=py38h2e5f0a9_0
numpy=1.21.2=py38h4b4dc7a_0
numpy-base=1.21.2=py38he0bd621_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.1=py38he9d5cce_1
pillow=9.0.1=py38hde71d04_0
pip=21.2.4=py38hecd8cb5_0
proto-plus=1.18.1=pyhd8ed1ab_0
protobuf=3.15.8=py38ha048514_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py38h50d1736_4
python=3.8.8=h88f2d9e_5
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.10.1=py_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.1.2=hca72f7f_1
requests=2.27.1=pyhd8ed1ab_0
rsa=4.8=pyhd8ed1ab_0
scikit-learn=1.0.2=py38hae1ba45_1
scipy=1.7.3=py38h8c7af03_0
setuptools=58.0.4=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=h707629a_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h7bc2e8c_0
tornado=6.1=py38h9ed2024_0
tqdm=4.63.0=pyhd3eb1b0_0
typing-extensions=4.1.1=hd8ed1ab_0
typing_extensions=4.1.1=pyha770c72_0
urllib3=1.26.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
yarl=1.7.2=py38h96a0964_1
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
